+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Yii Framework 2 authclient extension Change Log
 - Enh #341: OpenID Connect client now uses access token `'id_token'` claim for `getUserAttributes()` if `userinfo_endpoint` is not available (rhertogh)
 - Enh #342: OpenID Connect client support for JWT in `userinfo_endpoint` response (rhertogh)
 - Bug #344: Fix Facebook OAuth 400 error when latin characters are used in App name (pawelkania)
+- Enh #279: Add `AuthAction::$defaultClientId` and `AuthAction::getClientId()` (ditibal)
 
 
 2.2.11 August 09, 2021